Preparation of extractor panel for patient tests is functional
clicking the start test button on profile page instantiates a new
extractor prefab if there is currently zero instantiated or all are in
the middle of a test. Dialog panel is invoked which asks for
confirmation to add the patient representative prefab to the list of
samples on the extractor prefab. Otherwise the dialog is invoked in an
available extractor prefab that is still on standby. The next step is to
read input from device hardware to link up to the progress circle
(potentiometer?).