functional toggle switch and progress circle; incorporated FirstInitPanel script into FirstDevice script
Currently just using a key press (#1 on keyboard) to increase the
progress circle for testing. The progress circle only updates when the
toggle switch is set to ON which corresponds to the device runnning. A
dialog panel pops up when progress reaches 100% and click continue will
destroy the current device module after storing the names of the
finished samples into a list. Now a function should be created that
instantiates the centrifuge module when it detects that that list gets
filled.