Rest of device modules are functional (centrifuge,therm,imaging)
Takes finished samples from extraction and stores it in an intermediate
list. Update() is used to instantiate a centrifuge prefab + initiation
dialog whenever there are items in the list. After samples are selected
and confirmed, more samples can be added using the "+" button at the
top. After the switch is locked in and progress reaches 100%, the same
happens for Thermocycler and Imager. Next, functionality needs to be
coded for when Imager reaches 100%. When this happens, the viral load
reading should be recorded for each patient and the database be updated
accordingly.